add and implement new schemas
