+++ src/hooks/auth/change-password/usePostChangePassword.ts
@@ -1,4 +1,4 @@
-import { ChangePasswordSchema } from '@/lib/zod';
+
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
 import { SubmitHandler, useForm } from 'react-hook-form';
@@ -7,6 +7,7 @@ import { postChangePassword } from '@/server/auth/change-password/actions';
 import useGetToken from '@/hooks/common/useGetToken';
 import { z } from 'zod';
 import { ChangePassword } from '@/types';
+import { ChangePasswordSchema } from '@/schemas';
 
 type FormsFields = z.infer<typeof ChangePasswordSchema>;
