+++ src/hooks/auth/login/usePostLogin.ts
@@ -1,4 +1,5 @@
-import { LoginSchema } from '@/lib/zod';
+
+import { LoginSchema } from '@/schemas';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Console } from 'console';
 import { signIn } from 'next-auth/react';
