+++ src/hooks/auth/reset-password/usePostResetPasssword.ts
@@ -2,12 +2,12 @@ import { useMutation } from '@tanstack/react-query';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { postResetPassword } from '../../../server/auth/reset-password/actions';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { ResetPasswordSchema } from '@/lib/zod';
 import { useRouter } from 'next/navigation';
 import { signIn } from 'next-auth/react';
 import toast from 'react-hot-toast';
 import { z } from 'zod';
 import { ResetPassword } from '@/types';
+import { ResetPasswordSchema } from '@/schemas';
 
 interface Props {
   token: string;
